Backend/requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
art==6.4
asciimatics==1.15.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.2
click==8.1.8
cohere==5.16.1
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastavro==1.11.1
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.2
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
lxml==5.4.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy==1.15.0
mypy-extensions==1.0.0
openai==1.90.0
orjson==3.10.16
packaging==24.2
passlib==1.7.4
pillow==11.0.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.11.3
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.1
pyfiglet==1.0.2
Pygments==2.19.1
PyMuPDF==1.26.1
PyPDF2==3.0.1
pytest==8.3.3
python-docx==1.2.0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.3
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.2
tokenizers==0.21.2
tqdm==4.67.1
typer==0.15.2
types-requests==2.32.4.20250611
typing-inspection==0.4.0
typing_extensions==4.12.2
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3